+++ Changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASED]
+### Added
+ - Options to indicate bootstrap version manually.
+
 ### Fixed
  - Case with TextArea, when value is not string or JSON object.